refactor(phase5): extract learn commands to lib/commands/learn (52 lines, -179)
